requirements.txt
blobfile==3.0.0
clip==1.0
dearpygui==1.11.1
deepspeed==0.15.1
diffusers==0.30.3
imageio==2.34.2
kornia==0.7.3
matplotlib==3.6.3
mpi4py==4.0.0
numpy==1.24.1
opencv_python==4.10.0.84
packaging==24.1
pandas==2.0.3
Pillow==10.4.0
PyMCubes==0.1.6
pyrender==0.1.45
rich==13.8.1
scipy==1.14.1
setuptools==71.0.4
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
torch==1.13.1
torch_ema==0.3
torchvision==0.14.1
torchvision==0.19.0
tqdm==4.66.4
transformers==4.44.2
trimesh==4.4.3
xformers==0.0.28.post1
